# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Small Town Watchdog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entities:
    get:
      operationId: listEntities
      tags: [entities]
      summary: List all government entities
      parameters:
        - name: state
          in: query
          schema:
            type: string
          description: Filter by state code (e.g. LA, MS)
        - name: city
          in: query
          schema:
            type: string
          description: Filter by city name (case-insensitive)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"
    post:
      operationId: createEntity
      tags: [entities]
      summary: Create a government entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [entities]
      summary: Get a government entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          description: Not found
    patch:
      operationId: updateEntity
      tags: [entities]
      summary: Update a government entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          description: Not found
    delete:
      operationId: deleteEntity
      tags: [entities]
      summary: Delete a government entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List public records documents
      parameters:
        - name: entityId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: alertCategory
          in: query
          schema:
            type: string
        - name: redFlagLevel
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a public records document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /documents/{id}/summarize:
    post:
      operationId: summarizeDocument
      tags: [documents]
      summary: Generate AI plain-English summary for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
...
  /email-subscribers:
    post:
      operationId: createEmailSubscriber
      tags: [email-subscribers]
      summary: Opt in to neutral civic update emails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSubscriberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
  /email-subscribers/unsubscribe:
    post:
      operationId: unsubscribeEmail
      tags: [email-subscribers]
      summary: Unsubscribe an email from civic updates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailUnsubscribeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
...
components:
  schemas:
...
    Entity:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        state:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        documentCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    EntityInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        state:
          type: string
        city:
          type: string
        county:
          type: string
        location:
          type: string
        website:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    EntityPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        state:
          type: string
        city:
          type: string
        county:
          type: string
        location:
          type: string
        website:
          type: string
        description:
          type: string
        isActive:
          type: boolean
...
    Document:
      type: object
      required: [id, entityId, title, docType, year, createdAt]
      properties:
        id:
          type: integer
        entityId:
          type: integer
        entityName:
          type: ["string", "null"]
        sourceId:
          type: ["integer", "null"]
        title:
          type: string
        docType:
          type: string
        year:
          type: integer
        content:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        sourceName:
          type: ["string", "null"]
        pulledAt:
          type: ["string", "null"]
        plainSummary:
          type: ["string", "null"]
        eli12Summary:
          type: ["string", "null"]
        isAiGenerated:
          type: boolean
        alertCategory:
          type: ["string", "null"]
        redFlagLevel:
          type: ["string", "null"]
        amountInvolved:
          type: ["number", "null"]
        sourceStatus:
          type: string
          description: "valid | broken | missing | pending_review"
        lastVerifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DocumentInput:
      type: object
      required: [entityId, title, docType, year]
      properties:
        entityId:
          type: integer
        sourceId:
          type: integer
        title:
          type: string
          minLength: 1
        docType:
          type: string
        year:
          type: integer
        content:
          type: string
        sourceUrl:
          type: string
        sourceName:
          type: string
        plainSummary:
          type: string
        eli12Summary:
          type: string
        alertCategory:
          type: string
        redFlagLevel:
          type: string
        amountInvolved:
          type: number
    DocumentPatch:
      type: object
      properties:
        title:
          type: string
        docType:
          type: string
        year:
          type: integer
        content:
          type: string
        sourceUrl:
          type: string
        sourceName:
          type: string
        plainSummary:
          type: string
        eli12Summary:
          type: string
        alertCategory:
          type: string
        redFlagLevel:
          type: string
        amountInvolved:
          type: number
        sourceStatus:
          type: string
...
    EmailSubscriberInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        stateCode:
          type: string
        countyParish:
          type: string
        signupSource:
          type: string
    EmailUnsubscribeInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...